requirements.txt
discord
emoji
google-generativeai
pillow
